change type

+++ offsets.hpp
@@ -2,10 +2,10 @@
 
 namespace offsets
 {
-	uint64_t dwEntityList   = 0x4DFFEF4;
-	uint64_t dwForceAttack  = 0x322DCFC;
-	uint64_t dwLocalPlayer  = 0xDEA964;
-	uint64_t m_fFlags       = 0x104;
-	uint64_t m_iCrosshairId = 0x11838;
-	uint64_t m_iTeamNum     = 0xF4;
+	uintptr_t dwEntityList   = 0x4DFFEF4;
+	uintptr_t dwForceAttack  = 0x322DCFC;
+	uintptr_t dwLocalPlayer  = 0xDEA964;
+	uintptr_t m_fFlags       = 0x104;
+	uintptr_t m_iCrosshairId = 0x11838;
+	uintptr_t m_iTeamNum     = 0xF4;
 }
